⚡️ Move polls to to client-side tRPC
