+++ apps/web/src/components/member-selector.tsx
@@ -12,11 +12,10 @@ import { UsersIcon } from "lucide-react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useOptimistic, useTransition } from "react";
 import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
-import type { MemberDTO } from "@/features/space/member/types";
 import { Trans } from "@/i18n/client";
 
 interface MemberSelectorProps {
-  members: MemberDTO[];
+  members: { userId: string; name: string; image?: string }[];
 }
 
 export function MemberSelector({ members }: MemberSelectorProps) {
