+++ apps/web/src/trpc/routers/index.ts
@@ -6,6 +6,7 @@ import { auth } from "./auth";
 import { calendars } from "./calendars";
 import { events } from "./events";
 import { polls } from "./polls";
+import { space } from "./space";
 import { user } from "./user";
 
 export const appRouter = mergeRouters(
@@ -14,6 +15,7 @@ export const appRouter = mergeRouters(
     auth,
     events,
     polls,
+    space,
     user,
     calendars,
   }),
